Move print.hpp to a utility dir
